add check for dropping database when using test db

+++ src/test/java/ee/hein/shoppinglistserver/service/MongoTestService.java
@@ -15,7 +15,9 @@ public class MongoTestService extends MongoService {
     @Autowired
     public MongoTestService(MongoClient mongoClient, MongoProperties mongoProperties) {
         super(mongoClient, mongoProperties.getDatabase());
-        database.drop();
+        if (mongoProperties.getUri().contains("localhost") || mongoProperties.getUri().contains("127.0.0.1")) {
+            database.drop();
+        }
     }
 
     public void emptyKeyspace() {
